QA fronted 3
