+++ src/components/index.js
@@ -1,0 +1,5 @@
+import SliderContainer from './SliderContainer'
+import CommentsComponent from './commentsComponent'
+import ModalComponent from './modalComponent'
+
+export { SliderContainer, CommentsComponent, ModalComponent }
