+++ src/screens/ChoiceScreen/index.js
@@ -1,0 +1,3 @@
+import ChoiceScreen from './ChoiceScreen'
+
+export default ChoiceScreen
